Split `yas.go` into smaller logical files (#40)
